pyproject.toml
[tool.poetry]
name = "loss-basins"
version = "0.1.0"
description = ""
authors = ["Tamera Lanham <tamera.lanham@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
torch = "^1.11.0"
ipykernel = "^6.10.0"
jupyter = "^1.0.0"
numpy = "^1.22.3"
torchvision = "^0.12.0"
tqdm = "^4.63.1"
matplotlib = "^3.5.1"
pytorch-lightning = "^1.6.0"
dill = "^0.3.4"
google-cloud-storage = "^2.2.1"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
pytest = "^7.1.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
